Update plan.py
Updated plan.py to add in checks for the graduation month and year, to ensure that the graduation month and year must be after the current month and year. e.g. 01/2025 as graduation month and year will be valid if current month and year is 12/2024, but 01/2025 as current month and year, however, will no longer be valid